managerapp/requirements.txt
asgiref==3.7.2
automium==0.2.6
automium-web==0.1.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.38.0
colorama==0.4.6
cryptography==42.0.4
Django==5.0.2
django-ckeditor==6.7.1
django-debug-toolbar==4.3.0
django-js-asset==2.2.0
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Flask==3.0.2
idna==3.6
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
jwcrypto==1.5.4
legacy==0.1.7
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.1
